Move taskassignmentlog to its own context
